Disallow expanding repeated fields in ExpandProto
Previously, when ExpandProto was given a repeated field, it would return
the first entry in it. This is an unexpected (and, therefore, unwelcome)
behavior. Additionally, we now have ExpandRepeatedProto to deal with
repeated fields explicitly.

Therefore, support for repeated fields in ExpandProto is being
deprecated. Now, if you try expanding them, the ExpandProto constructor
throws an IllegalArgumentException.